+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- localStorage fallback for all values functions
 - Function to generate a UserScript header from an object
 - Function to delete a value with full typing support
 - Function to get an object with all available keys and values